+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/Accounts.java
@@ -1,0 +1,10 @@
+package com.mapbox.services.android.navigation.v5.navigation;
+
+import androidx.annotation.NonNull;
+
+interface Accounts {
+  @NonNull
+  String obtainSkuToken();
+
+  void onEndNavigation();
+}
