+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/DisableSku.java
@@ -1,0 +1,4 @@
+package com.mapbox.services.android.navigation.v5.navigation;
+
+interface DisableSku extends Accounts {
+}
