+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/DisableSkuManager.java
@@ -1,4 +1,17 @@
 package com.mapbox.services.android.navigation.v5.navigation;
 
-interface DisableSkuManager extends AccountsManager {
+import androidx.annotation.NonNull;
+
+class DisableSkuManager implements DisableSku {
+
+  @NonNull
+  @Override
+  public String obtainSkuToken() {
+    return "";
+  }
+
+  @Override
+  public void onEndNavigation() {
+
+  }
 }
