--- libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/DisableSkuManagerImpl.java
@@ -1,17 +1,0 @@
-package com.mapbox.services.android.navigation.v5.navigation;
-
-import androidx.annotation.NonNull;
-
-class DisableSkuManagerImpl implements DisableSkuManager {
-
-  @NonNull
-  @Override
-  public String obtainSkuToken() {
-    return "";
-  }
-
-  @Override
-  public void onEndNavigation() {
-
-  }
-}
