+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/MapboxNavigation.java
@@ -431,7 +431,7 @@ public class MapboxNavigation implements ServiceConnection {
       MapboxMetricsReporter.disable();
       navigationService.stopSelf();
       navigationEventDispatcher.onNavigationEvent(false);
-      AccountsManagerImpl.getInstance(applicationContext).endNavigation();
+      AccountsManager.getInstance(applicationContext).endNavigation();
     }
   }
