+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/Mau.java
@@ -1,0 +1,5 @@
+package com.mapbox.services.android.navigation.v5.navigation;
+
+interface Mau extends Accounts {
+
+}
