+++ libandroid-navigation/src/main/java/com/mapbox/services/android/navigation/v5/navigation/Trips.java
@@ -1,0 +1,5 @@
+package com.mapbox.services.android.navigation.v5.navigation;
+
+interface Trips extends Accounts {
+
+}
